requirements.txt
mkepub==1.2
Markdown==3.3.6
python-dotenv==1.0.0
beautifulsoup4==4.11.1